Refactor website data handling to use IWebsiteDetails and streamline site fetching logic
